Implementation of parallel computation of the inverse discrete Fourier transform. Small change in "fourier_transform_in_parallel" and "inverse_fourier_transform".
Implementation of parallel computation of inverse discrete Fourier transform:
    - The implementation is in the file "inverse_fourier_transform_in_parallel";
    - Implementation of parallel computation of inverse discrete Fourier transform on 8 cores. (If there are fewer or more cores, this is not a problem).

Changes in the "fourier_transform_in_parallel" function.
    - The data type for intervals has been changed.

Changes in the "inverse_fourier_transform" function.
    - One comment has been added.

+++ code/fourier_transform_in_parallel.py
@@ -54,7 +54,7 @@ def fourier_transform_in_parallel(path_to_signal = "../data/input_signal.wav"):
 
         # Creating calculation intervals for each core
         step = int(index_Nyquist_frequency/8)
-        interval = np.zeros(9, dtype=np.int16)
+        interval = np.zeros(9, dtype=np.uint32)
         interval[0] = 0
         for i in range(1, 8):
             interval[i] = interval[i-1] + step
